Correct windows dtype error on eq

+++ python/tests/test_datasetslice.py
@@ -9,7 +9,7 @@ from .common import assert_export
 
 class TestDatasetSlice:
     def setup_method(self):
-        var = sc.Variable(['x'], values=np.arange(5))
+        var = sc.Variable(['x'], values=np.arange(5, dtype=np.int64))
         self._d = sc.Dataset(data={'a': var, 'b': var}, coords={'x': var})
 
     def test_slice_with_range_datasetview_then_dataarrayview(self):
